Formatter

+++ lib/src/commands/Interactivity.dart
@@ -1,7 +1,11 @@
 part of nyxx.commands;
 
 /// Creates new poll, generates options and collects results. Returns `Map<Emoji, int` as result. [timeout] is set by default to 10 minutes
-Future<Map<Emoji, int>> createPoll(TextChannel channel, String title, Map<Emoji, String> options, {Duration timeout: const Duration(minutes: 10), String message: null, bool delete: false}) async {
+Future<Map<Emoji, int>> createPoll(
+    TextChannel channel, String title, Map<Emoji, String> options,
+    {Duration timeout: const Duration(minutes: 10),
+    String message: null,
+    bool delete: false}) async {
   StringBuffer buffer = new StringBuffer();
 
   buffer.writeln(title);
@@ -9,12 +13,10 @@ Future<Map<Emoji, int>> createPoll(TextChannel channel, String title, Map<Emoji,
     buffer.writeln("${k.format()} - $v");
   });
 
-  if(message != null)
-  buffer.writeln(message);
+  if (message != null) buffer.writeln(message);
 
   var msg = await channel.send(content: buffer.toString());
-  for (var emoji in options.keys)
-    await msg.createReaction(emoji);
+  for (var emoji in options.keys) await msg.createReaction(emoji);
 
   // Clean memory. Or just null it. Maybe GC will clean this.
   buffer = null;
@@ -22,14 +24,13 @@ Future<Map<Emoji, int>> createPoll(TextChannel channel, String title, Map<Emoji,
   var m = new Map<Emoji, int>();
   return new Future<Map<Emoji, int>>(() async {
     await for (var r in msg.onReactionAdded) {
-      if(m.containsKey(r.emoji))
-      m[r.emoji] = m[r.emoji] += 1;
+      if (m.containsKey(r.emoji))
+        m[r.emoji] = m[r.emoji] += 1;
       else
-      m[r.emoji] = 1;
+        m[r.emoji] = 1;
     }
   }).timeout(timeout, onTimeout: () async {
-    if(delete)
-    await msg.delete();
+    if (delete) await msg.delete();
 
     return m;
   });
@@ -60,7 +61,8 @@ class Pagination {
   Pagination.fromList(this.pages, this.channel);
 
   /// Paginates a list of Strings - each String is different page.
-  Future<Message> paginate({Duration timeout: const Duration(seconds: 30)}) async {
+  Future<Message> paginate(
+      {Duration timeout: const Duration(seconds: 30)}) async {
     var nextEmoji = EmojisUnicode.arrow_forward;
     var backEmoji = EmojisUnicode.arrow_backward;
     var firstEmoji = EmojisUnicode.track_previous;
@@ -74,26 +76,29 @@ class Pagination {
 
     new Future(() async {
       var currPage = 0;
-      var group = Util.merge([channel.client.onMessageReactionAdded, channel.client.onMessageReactionRemove]);
+      var group = Util.merge([
+        channel.client.onMessageReactionAdded,
+        channel.client.onMessageReactionRemove
+      ]);
 
-      await for(var event in group) {
-        if(msg.id != event.message.id) continue;
-        if(event.user.bot) continue;
+      await for (var event in group) {
+        if (msg.id != event.message.id) continue;
+        if (event.user.bot) continue;
 
         var emoji = event.emoji as UnicodeEmoji;
-        if(emoji.code == nextEmoji.encode()) {
-          if(currPage <= pages.length - 2) {
+        if (emoji.code == nextEmoji.encode()) {
+          if (currPage <= pages.length - 2) {
             ++currPage;
             await msg.edit(content: pages[currPage]);
           }
         } else if (emoji.code == backEmoji.encode()) {
-          if(currPage >= 1) {
+          if (currPage >= 1) {
             --currPage;
             await msg.edit(content: pages[currPage]);
           }
         } else if (emoji.code == firstEmoji.encode()) {
-            await msg.edit(content: pages.first);
-            currPage = 0;
+          await msg.edit(content: pages.first);
+          currPage = 0;
         } else if (emoji.code == lastEmoji.encode()) {
           await msg.edit(content: pages.last);
           currPage = pages.length;
